requirements.txt
fastapi
uvicorn
langchain
langchain-groq
python-dotenv
pydantic
sqlmodel
twilio
python-socketio[asgi]